Youtube_summarizer/requirements.txt
gradio==4.37.1
huggingface_hub==0.23.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
google-generativeai==0.7.2
tenacity==8.2.3
yt-dlp==2023.11.16
youtube-transcript-api==0.6.2
